# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BackEnd API",
        "title": "Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/order": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Order Movie",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie Name",
                        "name": "movie_cinema_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Quantity",
                        "name": "quantity",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Time",
                        "name": "time",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "description": "Seat",
                        "name": "seat[]",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.TaskResponse2"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.OrderBody"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/order/cinema": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Order Movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Choose Date ",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Choose Time",
                        "name": "time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Choose Location",
                        "name": "location",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.TaskResponse2"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.Cinema"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.TaskResponse2": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Cinema": {
            "type": "object",
            "properties": {
                "cinema_name": {
                    "type": "string",
                    "example": "Spiderman"
                },
                "date": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "location": {
                    "type": "string",
                    "example": "Bandung"
                },
                "time": {
                    "type": "string",
                    "example": "22:10:33"
                }
            }
        },
...
        "models.OrderBody": {
            "type": "object",
            "properties": {
                "cinema_name": {
                    "type": "string",
                    "example": "Hiflix"
                },
                "date": {
                    "type": "string",
                    "example": "2006-01-02"
                },
                "genre": {
                    "type": "string",
                    "example": "Action, Advanture"
                },
                "image_movie": {
                    "type": "string",
                    "example": "b00db012-1a27-43b3-895a-abd3f540362e.jpg"
                },
                "location": {
                    "type": "string",
                    "example": "Bandung"
                },
                "movie_cinema_id": {
                    "type": "integer",
                    "example": 1
                },
                "quantity": {
                    "type": "integer",
                    "example": 5
                },
                "seat": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "time": {
                    "type": "string",
                    "example": "22:10:33"
                },
                "title": {
                    "type": "string",
                    "example": "Spiderman"
                },
                "total_price": {
                    "type": "integer",
                    "example": 90000
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}